add new UI classes to the windows build and fix a couple UI issues

+++ src/PlayStateControl.cpp
@@ -13,7 +13,7 @@ PlayStateControl::PlayStateControl(IPlugBase* pPlug, IRECT rect, const IColor& b
 
 PlayState PlayStateControl::GetPlayState()
 {
-	return (PlayState)GetParam()->GetNonNormalized(mValue);
+	return (PlayState)((int)GetParam()->GetNonNormalized(mValue));
 }
 
 void PlayStateControl::SetPlayState(PlayState state)
